+++ tests/test_types.py
@@ -1,0 +1,56 @@
+import pytest
+
+import click
+
+
+@pytest.mark.parametrize(
+    ("type", "value", "expect"),
+    [
+        (click.IntRange(0, 5), "3", 3),
+        (click.IntRange(5), "5", 5),
+        (click.IntRange(5), "100", 100),
+        (click.IntRange(max=5), "5", 5),
+        (click.IntRange(max=5), "-100", -100),
+        (click.IntRange(0, clamp=True), "-1", 0),
+        (click.IntRange(max=5, clamp=True), "6", 5),
+        (click.IntRange(0, min_open=True, clamp=True), "0", 1),
+        (click.IntRange(max=5, max_open=True, clamp=True), "5", 4),
+        (click.FloatRange(0.5, 1.5), "1.2", 1.2),
+        (click.FloatRange(0.5, min_open=True), "0.51", 0.51),
+        (click.FloatRange(max=1.5, max_open=True), "1.49", 1.49),
+        (click.FloatRange(0.5, clamp=True), "-0.0", 0.5),
+        (click.FloatRange(max=1.5, clamp=True), "inf", 1.5),
+    ],
+)
+def test_range(type, value, expect):
+    assert type.convert(value, None, None) == expect
+
+
+@pytest.mark.parametrize(
+    ("type", "value", "expect"),
+    [
+        (click.IntRange(0, 5), "6", "6 is not in the range 0<=x<=5."),
+        (click.IntRange(5), "4", "4 is not in the range x>=5."),
+        (click.IntRange(max=5), "6", "6 is not in the range x<=5."),
+        (click.IntRange(0, 5, min_open=True), 0, "0<x<=5"),
+        (click.IntRange(0, 5, max_open=True), 5, "0<=x<5"),
+        (click.FloatRange(0.5, min_open=True), 0.5, "x>0.5"),
+        (click.FloatRange(max=1.5, max_open=True), 1.5, "x<1.5"),
+    ],
+)
+def test_range_fail(type, value, expect):
+    with pytest.raises(click.BadParameter) as exc_info:
+        type.convert(value, None, None)
+
+    assert expect in exc_info.value.message
+
+
+def test_float_range_no_clamp_open():
+    with pytest.raises(TypeError):
+        click.FloatRange(0, 1, max_open=True, clamp=True)
+
+    sneaky = click.FloatRange(0, 1, max_open=True)
+    sneaky.clamp = True
+
+    with pytest.raises(RuntimeError):
+        sneaky.convert("1.5", None, None)