+++ CHANGES.rst
@@ -18,9 +18,11 @@ Unreleased
     :issue:`1582`
 -   If validation fails for a prompt with ``hide_input=True``, the value
     is not shown in the error message. :issue:`1460`
--   An ``IntRange`` option shows the accepted range in its help text.
-    :issue:`1525`
--   ``FloatRange`` min and max can be exclusive. :issue:`1100`
+-   An ``IntRange`` or ``FloatRange`` option shows the accepted range in
+    its help text. :issue:`1525`, :pr:`1303`
+-   ``IntRange`` and ``FloatRange`` bounds can be open (``<``) instead
+    of closed (``<=``) by setting ``min_open`` and ``max_open``. Error
+    messages have changed to reflect this. :issue:`1100`
 -   An option defined with duplicate flag names (``"--foo/--foo"``)
     raises a ``ValueError``. :issue:`1465`
 -   ``echo()`` will not fail when using pytest's ``capsys`` fixture on